refactor: enhance database logic
- Wrong variable uid_chunks passed in any_not_previously_downloaded()
- Wrong value download_name returned in file_downloaded() when using
  time zone offsets
- Add assertion to ensure get_first_uid_from_uid_list() never receives
  any empty list of uids
- Return type annotation mismatch in get_uids_for_device()
- no_downloaded and no_thumbnails method names are misleading; rename
  as count_downloaded and count_thumbnails
- Correct type annotation for count_downloaded
- Change return type of md5_names() to a set
- Remove redundant index from devices table and replace it with
  index on files table
- Improve typing a notation for instance method get_uids() by using
  Typing.overload decorator
- Turn self.time_zone_offsets into a class attribute, avoiding redundant
  computation
- Exception Messages in add_downloaded_file and add_thumbnail are no
  longer lost
- Remove misplaced @retry decorator from _delete method
- Make add_thumbnail_rows handle large batches
- Convert deque buffer used when adding thumbnails to the main window UI
  to be a Python list
- Purge unused set_all_marked_as_unmarked method
- Convert InCache and ThumbnailRow to use typing.NamedTuple
- Apply Write Ahead Logging mode to downloaded_files.sqlite and
  thumbnail_cache.sqlite, increasing performance when one process writes
  and other processes read.
- Replace fragile NATURAL JOIN with explicit join
- When query in get_first_uid_from_uid_list exceeds maximum number of
  host parameters SQLite can handle, use temporary table to perform
  query
- Filter logged query length to limit parameter representation
- Don't mutate caller's proximity value list while building SQL
  where clause
- Query SQLite for its maximum number of host parameters

+++ raphodo/cache.py
@@ -29,6 +29,7 @@ For the fdo cache specs, see:
 http://specifications.freedesktop.org/thumbnail-spec/thumbnail-spec-latest.html
 """
 
+import contextlib
 import hashlib
 import logging
 import os
@@ -526,7 +527,8 @@ class ThumbnailCacheSql:
 
         path = os.path.join(self.cache_dir, in_cache.md5_name)
         if not os.path.exists(path):
-            self.thumb_db.delete_thumbnails([in_cache.md5_name])
+            with contextlib.suppress(sqlite3.DatabaseError):
+                self.thumb_db.delete_thumbnails([in_cache.md5_name])
             return self.not_found
 
         return GetThumbnailPath(
@@ -554,13 +556,13 @@ class ThumbnailCacheSql:
                 ):
                     os.remove(thumbnail)
                     deleted_thumbnails.append(name)
-            if len(deleted_thumbnails):
-                if self.thumb_db.cache_exists():
+            if len(deleted_thumbnails) and self.thumb_db.cache_exists():
+                with contextlib.suppress(sqlite3.DatabaseError):
                     self.thumb_db.delete_thumbnails(deleted_thumbnails)
-                logging.debug(
-                    f"Deleted {len(deleted_thumbnails)} thumbnail files that had not "
-                    f"been accessed for {days} or more days"
-                )
+                    logging.debug(
+                        f"Deleted {len(deleted_thumbnails)} thumbnail files that had not "
+                        f"been accessed for {days} or more days"
+                    )
 
     def purge_cache(self) -> None:
         """
@@ -568,17 +570,18 @@ class ThumbnailCacheSql:
         directory
         """
         if self.valid and self.cache_dir is not None and os.path.isdir(self.cache_dir):
-            # Delete the sqlite3 database too
+            # The sqlite3 database is located within this directory
+            # It will also be deleted.
             shutil.rmtree(self.cache_dir)
 
-    def no_thumbnails(self) -> int:
+    def count_thumbnails(self) -> int:
         """
         :return: how many thumbnails there are in the thumbnail database
         """
 
         if not self.valid:
             return 0
-        return self.thumb_db.no_thumbnails()
+        return self.thumb_db.count_thumbnails()
 
     def cache_size(self) -> int:
         """
@@ -612,13 +615,14 @@ class ThumbnailCacheSql:
         """
 
         rows = self.thumb_db.md5_names()
-        rows = {row[0] for row in rows}
         cwd = os.getcwd()
+        assert self.cache_dir is not None
         os.chdir(self.cache_dir)
 
         to_delete_from_db = {md5 for md5 in rows if not os.path.exists(md5)}
         if len(to_delete_from_db):
-            self.thumb_db.delete_thumbnails(list(to_delete_from_db))
+            with contextlib.suppress(sqlite3.DatabaseError):
+                self.thumb_db.delete_thumbnails(list(to_delete_from_db))
 
         md5s = {md5 for md5 in os.listdir(".")} - {self.thumb_db.db_fs_name()}
         to_delete_from_fs = md5s - rows
